% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{boston}
\alias{boston}
\title{Housing Values in Suburbs of Boston}
\format{A data.frame with 506 rows and 14 variables.
\itemize{
 \item \code{crim}: per capita crime rate by town.
 \item \code{zn}: proportion of residential land zoned for
 lots over 25,000 sq.ft.
  \item \code{indus}: proportion of non-retail business acres per town.
	\item \code{chas}: Charles River dummy variable (1 if tract bounds river;
  0 otherwise).
	\item \code{nox}: nitrogen oxides concentration (parts per 10 million).
	\item \code{rm}: average number of rooms per dwelling.
	\item \code{age}: proportion of owner-occupied units built prior to 1940.
	\item \code{dis}: weighted mean of distances to five Boston
  employment centres.
	\item \code{rad}: index of accessibility to radial highways.
	\item \code{tax}: full-value property-tax rate per $10,000.
 \item \code{ptratio}: pupil-teacher ratio by town.
	\item \code{black}: 1000(Bk - 0.63)^2 where Bk is the proportion of
  blacks by town.
	\item \code{lstat}: lower status of the population (percent).
	\item \code{medv}: median value of owner-occupied homes in $1000s.
}}
\source{
\itemize{
  \item Harrison, D. and Rubinfeld, D.L. (1978) Hedonic prices and the
  demand for clean air. \emph{J. Environ. Economics and Management} \bold{5}, 81--102.
  \item Belsley D.A., Kuh, E.  and Welsch, R.E. (1980) \emph{Regression
  Diagnostics. Identifying Influential Data and Sources of Collinearity.} New York: Wiley.
 }
}
\usage{
boston
}
\description{
This dataset is intended to be used in the modellatoR demos
for regression.
}
\author{
This dataset is part of the MASS package and has copyright 1994-9
W. N. Venables and B. D. Ripley.
}
\keyword{datasets}
\keyword{modellatoR,}

