% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{cancer_w}
\alias{cancer_w}
\title{BNG Breast Cancer Wisconsin}
\format{A data.frame with 39366 rows and 10 variables. Each row contains 9
predictors and 1 response (\code{Class}).
\itemize{
 \item \code{Clump_Thickness}: num  4.8 5.59 5.17 8.21 1 ...
	\item \code{Cell_Size_Uniformity}: num  1 1 1 4.56 1 ...
	\item \code{Cell_Shape_Uniformity}: num  1 1 1 5.43 1 ...
	\item \code{Marginal_Adhesion}: num  1 1 1 3.97 1 ...
	\item \code{Single_Epi_Cell_Size}: num  2 2 2 3 2 2 1 2 2 2 ...
	\item \code{Bare_Nuclei}: num  1 1 1 6.74 1 ...
	\item \code{Bland_Chromatin}: num  2 2 3 9.16 1 ...
	\item \code{Normal_Nucleoli}: num  1 1 1 10 1 ...
	\item \code{Mitoses}: num  1 1 1 1 1 1 1 1 1 1 ...
	\item \code{Class}: Factor w/ 2 levels "benign","malignant": 1 1 1 2 1 1 1 1 1 1 ...
}}
\source{
openML - \href{http://www.openml.org/d/251}{dataset 251}
}
\usage{
cancer_w
}
\description{
This dataset is intended to be used in the modellatoR demos
for classification. It's a larger version of the Breast Cancer Wisconsin
dataset made via a Bayesian Network trained on the original dataset and
then using it to create pseudo instances.
}
\author{
This extended version was made by Geoffrey Holmes,
Bernhard Pfahringer, Jan van Rijn and Joaquin Vanschoren.

The original dataset was made by Dr. WIlliam H. Wolberg and donated by
Olvi Mangasarian and David W. Aha.
}
\keyword{datasets}
\keyword{modellatoR,}

