% Generated by roxygen2 (4.0.2): do not edit by hand
\name{create_project}
\alias{create_project}
\title{Creates a folder structure/architecture for the project.}
\usage{
create_project(working_dir, project_name, project_minimal)
}
\arguments{
\item{working_dir}{is the path of the root folder where the
project's folder is going to be created.}

\item{project_name}{is the name of the project.}

\item{project_minimal,}{if set to TRUE, sets up a minimal folder structure.}
}
\value{
A folder with the project's folder structure and all the tools
delivered by the package `ProjectTemplate`.
}
\description{
Creates a folder structure/architecture for the project.
}

