% Generated by roxygen2 (4.0.1): do not edit by hand
\name{fit_random_forest}
\alias{fit_random_forest}
\title{Fits a Random Forest Model to the data.}
\usage{
fit_random_forest(data, params)
}
\arguments{
\item{data}{is a data.frame or data.table containing the data.}

\item{params}{is a list containing the arguments necessary for training.}
}
\value{
This function returns the chosen model.
}
\description{
Fits a Random Forest Model to the data.
}

