% Generated by roxygen2 (4.0.1): do not edit by hand
\name{setup_model}
\alias{setup_model}
\title{Creates the setup of the project.}
\usage{
setup_model(data, method_id, out_var)
}
\arguments{
\item{data}{is a data.frame or data.table containing the data.}

\item{method_id}{is a method for training. Needs to be one
of the following: 'dt', 'rf', 'glm', 'nn'. These acronyms correspond to:
\itemize{
  \item dt: Decision Trees Model
  \item rf: Random Forest Model
  \item glm: Generalized Liner Model
  \item nn: Neural Network Model
}}

\item{out_var}{is the target variable that the model tries to estimate.}
}
\value{
Writes the model's params to a config file at `config/params.R`.
}
\description{
Creates the setup of the project.
}

