% Generated by roxygen2 (4.0.2): do not edit by hand
\name{test_model}
\alias{test_model}
\title{Tests a model.}
\usage{
test_model(model, params, testset)
}
\arguments{
\item{model}{is a model computed with train_model.}

\item{params}{is the list of arguments created by setup_model.}

\item{testset}{is the data.frame that contains the data
that we want to evaluate.}
}
\value{
Returns a data.frame that contains the real
values of the target variable and the model estimations.
}
\description{
Tests a model.
}

