% Generated by roxygen2 (4.0.1): do not edit by hand
\name{train_model}
\alias{train_model}
\title{Fits the chosen model to the data.}
\usage{
train_model(data, params)
}
\arguments{
\item{data}{is a data.frame or data.table containing the data.}

\item{params}{is a list containing the arguments necessary for training.}
}
\value{
This function returns the chosen model.
}
\description{
Fits the chosen model to the data.
}

