% Generated by roxygen2 (4.0.2): do not edit by hand
\name{fit_random_forest}
\alias{fit_random_forest}
\title{Fits a Random Forest Model to the data.}
\usage{
fit_random_forest(trainset, testset, params)
}
\arguments{
\item{trainset}{is a data.frame or data.table containing the train data.}

\item{testset}{is a data.frame or data.table containing the test data.}
}
\value{
This function returns the chosen model.
}
\description{
Fits a Random Forest Model to the data.
}

