% Generated by roxygen2 (4.0.2): do not edit by hand
\name{train_model}
\alias{train_model}
\title{Fits the chosen model to the data.}
\usage{
train_model(trainset, testset, params)
}
\arguments{
\item{trainset}{is a data.frame or data.table containing the train data.}

\item{testset}{is a data.frame or data.table containing the test data.}

\item{params}{is a list containing the arguments necessary for training.}
}
\value{
This function returns the chosen model.
}
\description{
Fits the chosen model to the data.
}

